Remove $ from ONE file
In most places within this project $ is _not_ used to indicate
commands at the command prompt,
so I am bringing this file into compliance with that convention.
There are some places where $ is used that I think it may still make
sense (in order to distinguish between commands and output). I'll
discuss that with Daniele before committing.